chore(suite-native): decouple autoconnect hook
chore(suite-native): update suite-native/thp deps
