+++ suite-native/thp/src/index.ts
@@ -2,4 +2,4 @@ export * from './components/ThpCodeEntryScreenContent';
 export * from './components/ThpPairingInfoHelpButton';
 export * from './components/ThpPairingInfoScreenContent';
 export * from './components/ThpPairingSuccessScreenContent';
-export * from './hooks/useThpAutoconnectAlerts';
+export * from './hooks/useThpAutoconnectActions';
