+++ suite-native/thp/tsconfig.json
@@ -9,8 +9,6 @@
         { "path": "../alerts" },
         { "path": "../atoms" },
         { "path": "../intl" },
-        { "path": "../navigation" },
-        { "path": "../toasts" },
         { "path": "../../packages/connect" },
         { "path": "../../packages/styles" }
     ]
